move bin scripts into bin, apis into atomic-red-team
