+++ LICENSE.txt
@@ -1,7 +1,7 @@
 
 The MIT License
 
-Copyright (c) 2016 Red Canary, Inc.
+Copyright (c) 2018 Red Canary, Inc.
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
