+++ atomic-red-team/atomic_doc_template.md.erb
