+++ atomic-red-team/atomic_test_template.yaml
@@ -20,7 +20,7 @@ atomic_tests:
       type: todo
       default: TODO
 
-  executors:
-    name: TODO
+  executor:
+    name: command_prompt
     command: |
       TODO
