+++ atomic-red-team/attack_api.rb
