+++ atomic-red-team/spec.yaml
