+++ bin/new-atomic.rb
